simple_bert_classification/requirements.txt
torch
pytorch-transformers
pytreebank
loguru
tqdm